Make each variant case and record field unique
Previously, we were assuming that if two variant cases
have the same tag and have the same type, then arguments
of both cases would be of the same time. I can't think
of any situation where this is not true, but who knows...

Instead, we now use a witness to generate new type equalities
when new variant case and record fields are generated. And
we use that witness to ensure that the arguments of the
variant cases are equal.

I am not sure this is really necessary but this removes the
only uses of Obj.magic in the code, which is a reason good
enough to do it. Yay!

Thanks a lots to @let-def for providing most of the patch
and the patient explanations.